+++ src/comment/hooks/useReplies.ts
@@ -1,7 +1,6 @@
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import { getBlockedByUsers } from '@/user/api/user';
 import { fetchRepliesOnce } from '@/comment/api/reply';
-import { addReplyToComment } from '@/comment/utils/commentUtils';
 import { Reply } from '@/comment/model/Reply';
 import { useAuth } from '@/shared/hooks/useAuth';
 
@@ -19,30 +18,3 @@ export function useReplies(boardId: string, postId: string, commentId: string) {
 
   return { replies };
 }
-
-export function useAddReply(boardId: string, postId: string, commentId: string) {
-  const { currentUser } = useAuth();
-  const userId = currentUser?.uid;
-  const queryClient = useQueryClient();
-
-  return useMutation(
-    async (content: string) => {
-      if (!currentUser) throw new Error('로그인이 필요합니다.');
-      return addReplyToComment(
-        boardId,
-        postId,
-        commentId,
-        content,
-        currentUser.uid,
-        currentUser.displayName,
-        currentUser.photoURL,
-      );
-    },
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries({ queryKey: ['replies', boardId, postId, commentId, userId] });
-        queryClient.invalidateQueries({ queryKey: ['replyCount', boardId, postId, commentId, userId] });
-      },
-    }
-  );
-}
